.verdent/verdent-projects/Arviva OS/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
uvloop==0.20.0
httptools==0.6.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
pgvector==0.3.3
redis[hiredis]==5.1.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
numpy==1.26.4
tenacity==9.0.0
psutil>=5.9.8
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
gitpython==3.1.43
structlog==24.4.0
prometheus-client==0.21.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
grpcio==1.67.0
